requirements.txt
fastapi
fastapi[standard]
uvicorn
trafilatura
pydantic
python-dotenv
